\name{data}
\alias{data}
\docType{data}
\title{
Simulated Data with Ordinal Response
}
\description{
A simulated dataset for testing ordinal classification methodologies. The first columnn (V1) is associated with the ordinal response: for class 0 V1 is distributed as N(0,1), for class 1 V1 is distributed as N(1.5,1), and for class 2 V1 is distributed as N(3,1). All other predictor variables are multivariable normally distributed with mean vector 0 and variance-covariance matrix I. The Bayes Error associated with this dataset is 0.302.
}
\usage{data(data)}
\format{
  A data frame with 90 observations on the following 1001 variables. The first 1000 variables are predictors, the last is the ordinal response.
  \describe{
    \item{\code{V1}}{normally distributed within each class 0, 1, and 2 as N(0,1), N(1.5,1), and N(3,1), respectively}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
    \item{\code{V13}}{a numeric vector}
    \item{\code{V14}}{a numeric vector}
    \item{\code{V15}}{a numeric vector}
    \item{\code{V16}}{a numeric vector}
    \item{\code{V17}}{a numeric vector}
    \item{\code{V18}}{a numeric vector}
    \item{\code{V19}}{a numeric vector}
    \item{\code{V20}}{a numeric vector}
    \item{\code{V21}}{a numeric vector}
    \item{\code{V22}}{a numeric vector}
    \item{\code{V23}}{a numeric vector}
    \item{\code{V24}}{a numeric vector}
    \item{\code{V25}}{a numeric vector}
    \item{\code{V26}}{a numeric vector}
    \item{\code{V27}}{a numeric vector}
    \item{\code{V28}}{a numeric vector}
    \item{\code{V29}}{a numeric vector}
    \item{\code{V30}}{a numeric vector}
    \item{\code{V31}}{a numeric vector}
    \item{\code{V32}}{a numeric vector}
    \item{\code{V33}}{a numeric vector}
    \item{\code{V34}}{a numeric vector}
    \item{\code{V35}}{a numeric vector}
    \item{\code{V36}}{a numeric vector}
    \item{\code{V37}}{a numeric vector}
    \item{\code{V38}}{a numeric vector}
    \item{\code{V39}}{a numeric vector}
    \item{\code{V40}}{a numeric vector}
    \item{\code{V41}}{a numeric vector}
    \item{\code{V42}}{a numeric vector}
    \item{\code{V43}}{a numeric vector}
    \item{\code{V44}}{a numeric vector}
    \item{\code{V45}}{a numeric vector}
    \item{\code{V46}}{a numeric vector}
    \item{\code{V47}}{a numeric vector}
    \item{\code{V48}}{a numeric vector}
    \item{\code{V49}}{a numeric vector}
    \item{\code{V50}}{a numeric vector}
    \item{\code{V51}}{a numeric vector}
    \item{\code{V52}}{a numeric vector}
    \item{\code{V53}}{a numeric vector}
    \item{\code{V54}}{a numeric vector}
    \item{\code{V55}}{a numeric vector}
    \item{\code{V56}}{a numeric vector}
    \item{\code{V57}}{a numeric vector}
    \item{\code{V58}}{a numeric vector}
    \item{\code{V59}}{a numeric vector}
    \item{\code{V60}}{a numeric vector}
    \item{\code{V61}}{a numeric vector}
    \item{\code{V62}}{a numeric vector}
    \item{\code{V63}}{a numeric vector}
    \item{\code{V64}}{a numeric vector}
    \item{\code{V65}}{a numeric vector}
    \item{\code{V66}}{a numeric vector}
    \item{\code{V67}}{a numeric vector}
    \item{\code{V68}}{a numeric vector}
    \item{\code{V69}}{a numeric vector}
    \item{\code{V70}}{a numeric vector}
    \item{\code{V71}}{a numeric vector}
    \item{\code{V72}}{a numeric vector}
    \item{\code{V73}}{a numeric vector}
    \item{\code{V74}}{a numeric vector}
    \item{\code{V75}}{a numeric vector}
    \item{\code{V76}}{a numeric vector}
    \item{\code{V77}}{a numeric vector}
    \item{\code{V78}}{a numeric vector}
    \item{\code{V79}}{a numeric vector}
    \item{\code{V80}}{a numeric vector}
    \item{\code{V81}}{a numeric vector}
    \item{\code{V82}}{a numeric vector}
    \item{\code{V83}}{a numeric vector}
    \item{\code{V84}}{a numeric vector}
    \item{\code{V85}}{a numeric vector}
    \item{\code{V86}}{a numeric vector}
    \item{\code{V87}}{a numeric vector}
    \item{\code{V88}}{a numeric vector}
    \item{\code{V89}}{a numeric vector}
    \item{\code{V90}}{a numeric vector}
    \item{\code{V91}}{a numeric vector}
    \item{\code{V92}}{a numeric vector}
    \item{\code{V93}}{a numeric vector}
    \item{\code{V94}}{a numeric vector}
    \item{\code{V95}}{a numeric vector}
    \item{\code{V96}}{a numeric vector}
    \item{\code{V97}}{a numeric vector}
    \item{\code{V98}}{a numeric vector}
    \item{\code{V99}}{a numeric vector}
    \item{\code{V100}}{a numeric vector}
    \item{\code{V101}}{a numeric vector}
    \item{\code{V102}}{a numeric vector}
    \item{\code{V103}}{a numeric vector}
    \item{\code{V104}}{a numeric vector}
    \item{\code{V105}}{a numeric vector}
    \item{\code{V106}}{a numeric vector}
    \item{\code{V107}}{a numeric vector}
    \item{\code{V108}}{a numeric vector}
    \item{\code{V109}}{a numeric vector}
    \item{\code{V110}}{a numeric vector}
    \item{\code{V111}}{a numeric vector}
    \item{\code{V112}}{a numeric vector}
    \item{\code{V113}}{a numeric vector}
    \item{\code{V114}}{a numeric vector}
    \item{\code{V115}}{a numeric vector}
    \item{\code{V116}}{a numeric vector}
    \item{\code{V117}}{a numeric vector}
    \item{\code{V118}}{a numeric vector}
    \item{\code{V119}}{a numeric vector}
    \item{\code{V120}}{a numeric vector}
    \item{\code{V121}}{a numeric vector}
    \item{\code{V122}}{a numeric vector}
    \item{\code{V123}}{a numeric vector}
    \item{\code{V124}}{a numeric vector}
    \item{\code{V125}}{a numeric vector}
    \item{\code{V126}}{a numeric vector}
    \item{\code{V127}}{a numeric vector}
    \item{\code{V128}}{a numeric vector}
    \item{\code{V129}}{a numeric vector}
    \item{\code{V130}}{a numeric vector}
    \item{\code{V131}}{a numeric vector}
    \item{\code{V132}}{a numeric vector}
    \item{\code{V133}}{a numeric vector}
    \item{\code{V134}}{a numeric vector}
    \item{\code{V135}}{a numeric vector}
    \item{\code{V136}}{a numeric vector}
    \item{\code{V137}}{a numeric vector}
    \item{\code{V138}}{a numeric vector}
    \item{\code{V139}}{a numeric vector}
    \item{\code{V140}}{a numeric vector}
    \item{\code{V141}}{a numeric vector}
    \item{\code{V142}}{a numeric vector}
    \item{\code{V143}}{a numeric vector}
    \item{\code{V144}}{a numeric vector}
    \item{\code{V145}}{a numeric vector}
    \item{\code{V146}}{a numeric vector}
    \item{\code{V147}}{a numeric vector}
    \item{\code{V148}}{a numeric vector}
    \item{\code{V149}}{a numeric vector}
    \item{\code{V150}}{a numeric vector}
    \item{\code{V151}}{a numeric vector}
    \item{\code{V152}}{a numeric vector}
    \item{\code{V153}}{a numeric vector}
    \item{\code{V154}}{a numeric vector}
    \item{\code{V155}}{a numeric vector}
    \item{\code{V156}}{a numeric vector}
    \item{\code{V157}}{a numeric vector}
    \item{\code{V158}}{a numeric vector}
    \item{\code{V159}}{a numeric vector}
    \item{\code{V160}}{a numeric vector}
    \item{\code{V161}}{a numeric vector}
    \item{\code{V162}}{a numeric vector}
    \item{\code{V163}}{a numeric vector}
    \item{\code{V164}}{a numeric vector}
    \item{\code{V165}}{a numeric vector}
    \item{\code{V166}}{a numeric vector}
    \item{\code{V167}}{a numeric vector}
    \item{\code{V168}}{a numeric vector}
    \item{\code{V169}}{a numeric vector}
    \item{\code{V170}}{a numeric vector}
    \item{\code{V171}}{a numeric vector}
    \item{\code{V172}}{a numeric vector}
    \item{\code{V173}}{a numeric vector}
    \item{\code{V174}}{a numeric vector}
    \item{\code{V175}}{a numeric vector}
    \item{\code{V176}}{a numeric vector}
    \item{\code{V177}}{a numeric vector}
    \item{\code{V178}}{a numeric vector}
    \item{\code{V179}}{a numeric vector}
    \item{\code{V180}}{a numeric vector}
    \item{\code{V181}}{a numeric vector}
    \item{\code{V182}}{a numeric vector}
    \item{\code{V183}}{a numeric vector}
    \item{\code{V184}}{a numeric vector}
    \item{\code{V185}}{a numeric vector}
    \item{\code{V186}}{a numeric vector}
    \item{\code{V187}}{a numeric vector}
    \item{\code{V188}}{a numeric vector}
    \item{\code{V189}}{a numeric vector}
    \item{\code{V190}}{a numeric vector}
    \item{\code{V191}}{a numeric vector}
    \item{\code{V192}}{a numeric vector}
    \item{\code{V193}}{a numeric vector}
    \item{\code{V194}}{a numeric vector}
    \item{\code{V195}}{a numeric vector}
    \item{\code{V196}}{a numeric vector}
    \item{\code{V197}}{a numeric vector}
    \item{\code{V198}}{a numeric vector}
    \item{\code{V199}}{a numeric vector}
    \item{\code{V200}}{a numeric vector}
    \item{\code{V201}}{a numeric vector}
    \item{\code{V202}}{a numeric vector}
    \item{\code{V203}}{a numeric vector}
    \item{\code{V204}}{a numeric vector}
    \item{\code{V205}}{a numeric vector}
    \item{\code{V206}}{a numeric vector}
    \item{\code{V207}}{a numeric vector}
    \item{\code{V208}}{a numeric vector}
    \item{\code{V209}}{a numeric vector}
    \item{\code{V210}}{a numeric vector}
    \item{\code{V211}}{a numeric vector}
    \item{\code{V212}}{a numeric vector}
    \item{\code{V213}}{a numeric vector}
    \item{\code{V214}}{a numeric vector}
    \item{\code{V215}}{a numeric vector}
    \item{\code{V216}}{a numeric vector}
    \item{\code{V217}}{a numeric vector}
    \item{\code{V218}}{a numeric vector}
    \item{\code{V219}}{a numeric vector}
    \item{\code{V220}}{a numeric vector}
    \item{\code{V221}}{a numeric vector}
    \item{\code{V222}}{a numeric vector}
    \item{\code{V223}}{a numeric vector}
    \item{\code{V224}}{a numeric vector}
    \item{\code{V225}}{a numeric vector}
    \item{\code{V226}}{a numeric vector}
    \item{\code{V227}}{a numeric vector}
    \item{\code{V228}}{a numeric vector}
    \item{\code{V229}}{a numeric vector}
    \item{\code{V230}}{a numeric vector}
    \item{\code{V231}}{a numeric vector}
    \item{\code{V232}}{a numeric vector}
    \item{\code{V233}}{a numeric vector}
    \item{\code{V234}}{a numeric vector}
    \item{\code{V235}}{a numeric vector}
    \item{\code{V236}}{a numeric vector}
    \item{\code{V237}}{a numeric vector}
    \item{\code{V238}}{a numeric vector}
    \item{\code{V239}}{a numeric vector}
    \item{\code{V240}}{a numeric vector}
    \item{\code{V241}}{a numeric vector}
    \item{\code{V242}}{a numeric vector}
    \item{\code{V243}}{a numeric vector}
    \item{\code{V244}}{a numeric vector}
    \item{\code{V245}}{a numeric vector}
    \item{\code{V246}}{a numeric vector}
    \item{\code{V247}}{a numeric vector}
    \item{\code{V248}}{a numeric vector}
    \item{\code{V249}}{a numeric vector}
    \item{\code{V250}}{a numeric vector}
    \item{\code{V251}}{a numeric vector}
    \item{\code{V252}}{a numeric vector}
    \item{\code{V253}}{a numeric vector}
    \item{\code{V254}}{a numeric vector}
    \item{\code{V255}}{a numeric vector}
    \item{\code{V256}}{a numeric vector}
    \item{\code{V257}}{a numeric vector}
    \item{\code{V258}}{a numeric vector}
    \item{\code{V259}}{a numeric vector}
    \item{\code{V260}}{a numeric vector}
    \item{\code{V261}}{a numeric vector}
    \item{\code{V262}}{a numeric vector}
    \item{\code{V263}}{a numeric vector}
    \item{\code{V264}}{a numeric vector}
    \item{\code{V265}}{a numeric vector}
    \item{\code{V266}}{a numeric vector}
    \item{\code{V267}}{a numeric vector}
    \item{\code{V268}}{a numeric vector}
    \item{\code{V269}}{a numeric vector}
    \item{\code{V270}}{a numeric vector}
    \item{\code{V271}}{a numeric vector}
    \item{\code{V272}}{a numeric vector}
    \item{\code{V273}}{a numeric vector}
    \item{\code{V274}}{a numeric vector}
    \item{\code{V275}}{a numeric vector}
    \item{\code{V276}}{a numeric vector}
    \item{\code{V277}}{a numeric vector}
    \item{\code{V278}}{a numeric vector}
    \item{\code{V279}}{a numeric vector}
    \item{\code{V280}}{a numeric vector}
    \item{\code{V281}}{a numeric vector}
    \item{\code{V282}}{a numeric vector}
    \item{\code{V283}}{a numeric vector}
    \item{\code{V284}}{a numeric vector}
    \item{\code{V285}}{a numeric vector}
    \item{\code{V286}}{a numeric vector}
    \item{\code{V287}}{a numeric vector}
    \item{\code{V288}}{a numeric vector}
    \item{\code{V289}}{a numeric vector}
    \item{\code{V290}}{a numeric vector}
    \item{\code{V291}}{a numeric vector}
    \item{\code{V292}}{a numeric vector}
    \item{\code{V293}}{a numeric vector}
    \item{\code{V294}}{a numeric vector}
    \item{\code{V295}}{a numeric vector}
    \item{\code{V296}}{a numeric vector}
    \item{\code{V297}}{a numeric vector}
    \item{\code{V298}}{a numeric vector}
    \item{\code{V299}}{a numeric vector}
    \item{\code{V300}}{a numeric vector}
    \item{\code{V301}}{a numeric vector}
    \item{\code{V302}}{a numeric vector}
    \item{\code{V303}}{a numeric vector}
    \item{\code{V304}}{a numeric vector}
    \item{\code{V305}}{a numeric vector}
    \item{\code{V306}}{a numeric vector}
    \item{\code{V307}}{a numeric vector}
    \item{\code{V308}}{a numeric vector}
    \item{\code{V309}}{a numeric vector}
    \item{\code{V310}}{a numeric vector}
    \item{\code{V311}}{a numeric vector}
    \item{\code{V312}}{a numeric vector}
    \item{\code{V313}}{a numeric vector}
    \item{\code{V314}}{a numeric vector}
    \item{\code{V315}}{a numeric vector}
    \item{\code{V316}}{a numeric vector}
    \item{\code{V317}}{a numeric vector}
    \item{\code{V318}}{a numeric vector}
    \item{\code{V319}}{a numeric vector}
    \item{\code{V320}}{a numeric vector}
    \item{\code{V321}}{a numeric vector}
    \item{\code{V322}}{a numeric vector}
    \item{\code{V323}}{a numeric vector}
    \item{\code{V324}}{a numeric vector}
    \item{\code{V325}}{a numeric vector}
    \item{\code{V326}}{a numeric vector}
    \item{\code{V327}}{a numeric vector}
    \item{\code{V328}}{a numeric vector}
    \item{\code{V329}}{a numeric vector}
    \item{\code{V330}}{a numeric vector}
    \item{\code{V331}}{a numeric vector}
    \item{\code{V332}}{a numeric vector}
    \item{\code{V333}}{a numeric vector}
    \item{\code{V334}}{a numeric vector}
    \item{\code{V335}}{a numeric vector}
    \item{\code{V336}}{a numeric vector}
    \item{\code{V337}}{a numeric vector}
    \item{\code{V338}}{a numeric vector}
    \item{\code{V339}}{a numeric vector}
    \item{\code{V340}}{a numeric vector}
    \item{\code{V341}}{a numeric vector}
    \item{\code{V342}}{a numeric vector}
    \item{\code{V343}}{a numeric vector}
    \item{\code{V344}}{a numeric vector}
    \item{\code{V345}}{a numeric vector}
    \item{\code{V346}}{a numeric vector}
    \item{\code{V347}}{a numeric vector}
    \item{\code{V348}}{a numeric vector}
    \item{\code{V349}}{a numeric vector}
    \item{\code{V350}}{a numeric vector}
    \item{\code{V351}}{a numeric vector}
    \item{\code{V352}}{a numeric vector}
    \item{\code{V353}}{a numeric vector}
    \item{\code{V354}}{a numeric vector}
    \item{\code{V355}}{a numeric vector}
    \item{\code{V356}}{a numeric vector}
    \item{\code{V357}}{a numeric vector}
    \item{\code{V358}}{a numeric vector}
    \item{\code{V359}}{a numeric vector}
    \item{\code{V360}}{a numeric vector}
    \item{\code{V361}}{a numeric vector}
    \item{\code{V362}}{a numeric vector}
    \item{\code{V363}}{a numeric vector}
    \item{\code{V364}}{a numeric vector}
    \item{\code{V365}}{a numeric vector}
    \item{\code{V366}}{a numeric vector}
    \item{\code{V367}}{a numeric vector}
    \item{\code{V368}}{a numeric vector}
    \item{\code{V369}}{a numeric vector}
    \item{\code{V370}}{a numeric vector}
    \item{\code{V371}}{a numeric vector}
    \item{\code{V372}}{a numeric vector}
    \item{\code{V373}}{a numeric vector}
    \item{\code{V374}}{a numeric vector}
    \item{\code{V375}}{a numeric vector}
    \item{\code{V376}}{a numeric vector}
    \item{\code{V377}}{a numeric vector}
    \item{\code{V378}}{a numeric vector}
    \item{\code{V379}}{a numeric vector}
    \item{\code{V380}}{a numeric vector}
    \item{\code{V381}}{a numeric vector}
    \item{\code{V382}}{a numeric vector}
    \item{\code{V383}}{a numeric vector}
    \item{\code{V384}}{a numeric vector}
    \item{\code{V385}}{a numeric vector}
    \item{\code{V386}}{a numeric vector}
    \item{\code{V387}}{a numeric vector}
    \item{\code{V388}}{a numeric vector}
    \item{\code{V389}}{a numeric vector}
    \item{\code{V390}}{a numeric vector}
    \item{\code{V391}}{a numeric vector}
    \item{\code{V392}}{a numeric vector}
    \item{\code{V393}}{a numeric vector}
    \item{\code{V394}}{a numeric vector}
    \item{\code{V395}}{a numeric vector}
    \item{\code{V396}}{a numeric vector}
    \item{\code{V397}}{a numeric vector}
    \item{\code{V398}}{a numeric vector}
    \item{\code{V399}}{a numeric vector}
    \item{\code{V400}}{a numeric vector}
    \item{\code{V401}}{a numeric vector}
    \item{\code{V402}}{a numeric vector}
    \item{\code{V403}}{a numeric vector}
    \item{\code{V404}}{a numeric vector}
    \item{\code{V405}}{a numeric vector}
    \item{\code{V406}}{a numeric vector}
    \item{\code{V407}}{a numeric vector}
    \item{\code{V408}}{a numeric vector}
    \item{\code{V409}}{a numeric vector}
    \item{\code{V410}}{a numeric vector}
    \item{\code{V411}}{a numeric vector}
    \item{\code{V412}}{a numeric vector}
    \item{\code{V413}}{a numeric vector}
    \item{\code{V414}}{a numeric vector}
    \item{\code{V415}}{a numeric vector}
    \item{\code{V416}}{a numeric vector}
    \item{\code{V417}}{a numeric vector}
    \item{\code{V418}}{a numeric vector}
    \item{\code{V419}}{a numeric vector}
    \item{\code{V420}}{a numeric vector}
    \item{\code{V421}}{a numeric vector}
    \item{\code{V422}}{a numeric vector}
    \item{\code{V423}}{a numeric vector}
    \item{\code{V424}}{a numeric vector}
    \item{\code{V425}}{a numeric vector}
    \item{\code{V426}}{a numeric vector}
    \item{\code{V427}}{a numeric vector}
    \item{\code{V428}}{a numeric vector}
    \item{\code{V429}}{a numeric vector}
    \item{\code{V430}}{a numeric vector}
    \item{\code{V431}}{a numeric vector}
    \item{\code{V432}}{a numeric vector}
    \item{\code{V433}}{a numeric vector}
    \item{\code{V434}}{a numeric vector}
    \item{\code{V435}}{a numeric vector}
    \item{\code{V436}}{a numeric vector}
    \item{\code{V437}}{a numeric vector}
    \item{\code{V438}}{a numeric vector}
    \item{\code{V439}}{a numeric vector}
    \item{\code{V440}}{a numeric vector}
    \item{\code{V441}}{a numeric vector}
    \item{\code{V442}}{a numeric vector}
    \item{\code{V443}}{a numeric vector}
    \item{\code{V444}}{a numeric vector}
    \item{\code{V445}}{a numeric vector}
    \item{\code{V446}}{a numeric vector}
    \item{\code{V447}}{a numeric vector}
    \item{\code{V448}}{a numeric vector}
    \item{\code{V449}}{a numeric vector}
    \item{\code{V450}}{a numeric vector}
    \item{\code{V451}}{a numeric vector}
    \item{\code{V452}}{a numeric vector}
    \item{\code{V453}}{a numeric vector}
    \item{\code{V454}}{a numeric vector}
    \item{\code{V455}}{a numeric vector}
    \item{\code{V456}}{a numeric vector}
    \item{\code{V457}}{a numeric vector}
    \item{\code{V458}}{a numeric vector}
    \item{\code{V459}}{a numeric vector}
    \item{\code{V460}}{a numeric vector}
    \item{\code{V461}}{a numeric vector}
    \item{\code{V462}}{a numeric vector}
    \item{\code{V463}}{a numeric vector}
    \item{\code{V464}}{a numeric vector}
    \item{\code{V465}}{a numeric vector}
    \item{\code{V466}}{a numeric vector}
    \item{\code{V467}}{a numeric vector}
    \item{\code{V468}}{a numeric vector}
    \item{\code{V469}}{a numeric vector}
    \item{\code{V470}}{a numeric vector}
    \item{\code{V471}}{a numeric vector}
    \item{\code{V472}}{a numeric vector}
    \item{\code{V473}}{a numeric vector}
    \item{\code{V474}}{a numeric vector}
    \item{\code{V475}}{a numeric vector}
    \item{\code{V476}}{a numeric vector}
    \item{\code{V477}}{a numeric vector}
    \item{\code{V478}}{a numeric vector}
    \item{\code{V479}}{a numeric vector}
    \item{\code{V480}}{a numeric vector}
    \item{\code{V481}}{a numeric vector}
    \item{\code{V482}}{a numeric vector}
    \item{\code{V483}}{a numeric vector}
    \item{\code{V484}}{a numeric vector}
    \item{\code{V485}}{a numeric vector}
    \item{\code{V486}}{a numeric vector}
    \item{\code{V487}}{a numeric vector}
    \item{\code{V488}}{a numeric vector}
    \item{\code{V489}}{a numeric vector}
    \item{\code{V490}}{a numeric vector}
    \item{\code{V491}}{a numeric vector}
    \item{\code{V492}}{a numeric vector}
    \item{\code{V493}}{a numeric vector}
    \item{\code{V494}}{a numeric vector}
    \item{\code{V495}}{a numeric vector}
    \item{\code{V496}}{a numeric vector}
    \item{\code{V497}}{a numeric vector}
    \item{\code{V498}}{a numeric vector}
    \item{\code{V499}}{a numeric vector}
    \item{\code{V500}}{a numeric vector}
    \item{\code{V501}}{a numeric vector}
    \item{\code{V502}}{a numeric vector}
    \item{\code{V503}}{a numeric vector}
    \item{\code{V504}}{a numeric vector}
    \item{\code{V505}}{a numeric vector}
    \item{\code{V506}}{a numeric vector}
    \item{\code{V507}}{a numeric vector}
    \item{\code{V508}}{a numeric vector}
    \item{\code{V509}}{a numeric vector}
    \item{\code{V510}}{a numeric vector}
    \item{\code{V511}}{a numeric vector}
    \item{\code{V512}}{a numeric vector}
    \item{\code{V513}}{a numeric vector}
    \item{\code{V514}}{a numeric vector}
    \item{\code{V515}}{a numeric vector}
    \item{\code{V516}}{a numeric vector}
    \item{\code{V517}}{a numeric vector}
    \item{\code{V518}}{a numeric vector}
    \item{\code{V519}}{a numeric vector}
    \item{\code{V520}}{a numeric vector}
    \item{\code{V521}}{a numeric vector}
    \item{\code{V522}}{a numeric vector}
    \item{\code{V523}}{a numeric vector}
    \item{\code{V524}}{a numeric vector}
    \item{\code{V525}}{a numeric vector}
    \item{\code{V526}}{a numeric vector}
    \item{\code{V527}}{a numeric vector}
    \item{\code{V528}}{a numeric vector}
    \item{\code{V529}}{a numeric vector}
    \item{\code{V530}}{a numeric vector}
    \item{\code{V531}}{a numeric vector}
    \item{\code{V532}}{a numeric vector}
    \item{\code{V533}}{a numeric vector}
    \item{\code{V534}}{a numeric vector}
    \item{\code{V535}}{a numeric vector}
    \item{\code{V536}}{a numeric vector}
    \item{\code{V537}}{a numeric vector}
    \item{\code{V538}}{a numeric vector}
    \item{\code{V539}}{a numeric vector}
    \item{\code{V540}}{a numeric vector}
    \item{\code{V541}}{a numeric vector}
    \item{\code{V542}}{a numeric vector}
    \item{\code{V543}}{a numeric vector}
    \item{\code{V544}}{a numeric vector}
    \item{\code{V545}}{a numeric vector}
    \item{\code{V546}}{a numeric vector}
    \item{\code{V547}}{a numeric vector}
    \item{\code{V548}}{a numeric vector}
    \item{\code{V549}}{a numeric vector}
    \item{\code{V550}}{a numeric vector}
    \item{\code{V551}}{a numeric vector}
    \item{\code{V552}}{a numeric vector}
    \item{\code{V553}}{a numeric vector}
    \item{\code{V554}}{a numeric vector}
    \item{\code{V555}}{a numeric vector}
    \item{\code{V556}}{a numeric vector}
    \item{\code{V557}}{a numeric vector}
    \item{\code{V558}}{a numeric vector}
    \item{\code{V559}}{a numeric vector}
    \item{\code{V560}}{a numeric vector}
    \item{\code{V561}}{a numeric vector}
    \item{\code{V562}}{a numeric vector}
    \item{\code{V563}}{a numeric vector}
    \item{\code{V564}}{a numeric vector}
    \item{\code{V565}}{a numeric vector}
    \item{\code{V566}}{a numeric vector}
    \item{\code{V567}}{a numeric vector}
    \item{\code{V568}}{a numeric vector}
    \item{\code{V569}}{a numeric vector}
    \item{\code{V570}}{a numeric vector}
    \item{\code{V571}}{a numeric vector}
    \item{\code{V572}}{a numeric vector}
    \item{\code{V573}}{a numeric vector}
    \item{\code{V574}}{a numeric vector}
    \item{\code{V575}}{a numeric vector}
    \item{\code{V576}}{a numeric vector}
    \item{\code{V577}}{a numeric vector}
    \item{\code{V578}}{a numeric vector}
    \item{\code{V579}}{a numeric vector}
    \item{\code{V580}}{a numeric vector}
    \item{\code{V581}}{a numeric vector}
    \item{\code{V582}}{a numeric vector}
    \item{\code{V583}}{a numeric vector}
    \item{\code{V584}}{a numeric vector}
    \item{\code{V585}}{a numeric vector}
    \item{\code{V586}}{a numeric vector}
    \item{\code{V587}}{a numeric vector}
    \item{\code{V588}}{a numeric vector}
    \item{\code{V589}}{a numeric vector}
    \item{\code{V590}}{a numeric vector}
    \item{\code{V591}}{a numeric vector}
    \item{\code{V592}}{a numeric vector}
    \item{\code{V593}}{a numeric vector}
    \item{\code{V594}}{a numeric vector}
    \item{\code{V595}}{a numeric vector}
    \item{\code{V596}}{a numeric vector}
    \item{\code{V597}}{a numeric vector}
    \item{\code{V598}}{a numeric vector}
    \item{\code{V599}}{a numeric vector}
    \item{\code{V600}}{a numeric vector}
    \item{\code{V601}}{a numeric vector}
    \item{\code{V602}}{a numeric vector}
    \item{\code{V603}}{a numeric vector}
    \item{\code{V604}}{a numeric vector}
    \item{\code{V605}}{a numeric vector}
    \item{\code{V606}}{a numeric vector}
    \item{\code{V607}}{a numeric vector}
    \item{\code{V608}}{a numeric vector}
    \item{\code{V609}}{a numeric vector}
    \item{\code{V610}}{a numeric vector}
    \item{\code{V611}}{a numeric vector}
    \item{\code{V612}}{a numeric vector}
    \item{\code{V613}}{a numeric vector}
    \item{\code{V614}}{a numeric vector}
    \item{\code{V615}}{a numeric vector}
    \item{\code{V616}}{a numeric vector}
    \item{\code{V617}}{a numeric vector}
    \item{\code{V618}}{a numeric vector}
    \item{\code{V619}}{a numeric vector}
    \item{\code{V620}}{a numeric vector}
    \item{\code{V621}}{a numeric vector}
    \item{\code{V622}}{a numeric vector}
    \item{\code{V623}}{a numeric vector}
    \item{\code{V624}}{a numeric vector}
    \item{\code{V625}}{a numeric vector}
    \item{\code{V626}}{a numeric vector}
    \item{\code{V627}}{a numeric vector}
    \item{\code{V628}}{a numeric vector}
    \item{\code{V629}}{a numeric vector}
    \item{\code{V630}}{a numeric vector}
    \item{\code{V631}}{a numeric vector}
    \item{\code{V632}}{a numeric vector}
    \item{\code{V633}}{a numeric vector}
    \item{\code{V634}}{a numeric vector}
    \item{\code{V635}}{a numeric vector}
    \item{\code{V636}}{a numeric vector}
    \item{\code{V637}}{a numeric vector}
    \item{\code{V638}}{a numeric vector}
    \item{\code{V639}}{a numeric vector}
    \item{\code{V640}}{a numeric vector}
    \item{\code{V641}}{a numeric vector}
    \item{\code{V642}}{a numeric vector}
    \item{\code{V643}}{a numeric vector}
    \item{\code{V644}}{a numeric vector}
    \item{\code{V645}}{a numeric vector}
    \item{\code{V646}}{a numeric vector}
    \item{\code{V647}}{a numeric vector}
    \item{\code{V648}}{a numeric vector}
    \item{\code{V649}}{a numeric vector}
    \item{\code{V650}}{a numeric vector}
    \item{\code{V651}}{a numeric vector}
    \item{\code{V652}}{a numeric vector}
    \item{\code{V653}}{a numeric vector}
    \item{\code{V654}}{a numeric vector}
    \item{\code{V655}}{a numeric vector}
    \item{\code{V656}}{a numeric vector}
    \item{\code{V657}}{a numeric vector}
    \item{\code{V658}}{a numeric vector}
    \item{\code{V659}}{a numeric vector}
    \item{\code{V660}}{a numeric vector}
    \item{\code{V661}}{a numeric vector}
    \item{\code{V662}}{a numeric vector}
    \item{\code{V663}}{a numeric vector}
    \item{\code{V664}}{a numeric vector}
    \item{\code{V665}}{a numeric vector}
    \item{\code{V666}}{a numeric vector}
    \item{\code{V667}}{a numeric vector}
    \item{\code{V668}}{a numeric vector}
    \item{\code{V669}}{a numeric vector}
    \item{\code{V670}}{a numeric vector}
    \item{\code{V671}}{a numeric vector}
    \item{\code{V672}}{a numeric vector}
    \item{\code{V673}}{a numeric vector}
    \item{\code{V674}}{a numeric vector}
    \item{\code{V675}}{a numeric vector}
    \item{\code{V676}}{a numeric vector}
    \item{\code{V677}}{a numeric vector}
    \item{\code{V678}}{a numeric vector}
    \item{\code{V679}}{a numeric vector}
    \item{\code{V680}}{a numeric vector}
    \item{\code{V681}}{a numeric vector}
    \item{\code{V682}}{a numeric vector}
    \item{\code{V683}}{a numeric vector}
    \item{\code{V684}}{a numeric vector}
    \item{\code{V685}}{a numeric vector}
    \item{\code{V686}}{a numeric vector}
    \item{\code{V687}}{a numeric vector}
    \item{\code{V688}}{a numeric vector}
    \item{\code{V689}}{a numeric vector}
    \item{\code{V690}}{a numeric vector}
    \item{\code{V691}}{a numeric vector}
    \item{\code{V692}}{a numeric vector}
    \item{\code{V693}}{a numeric vector}
    \item{\code{V694}}{a numeric vector}
    \item{\code{V695}}{a numeric vector}
    \item{\code{V696}}{a numeric vector}
    \item{\code{V697}}{a numeric vector}
    \item{\code{V698}}{a numeric vector}
    \item{\code{V699}}{a numeric vector}
    \item{\code{V700}}{a numeric vector}
    \item{\code{V701}}{a numeric vector}
    \item{\code{V702}}{a numeric vector}
    \item{\code{V703}}{a numeric vector}
    \item{\code{V704}}{a numeric vector}
    \item{\code{V705}}{a numeric vector}
    \item{\code{V706}}{a numeric vector}
    \item{\code{V707}}{a numeric vector}
    \item{\code{V708}}{a numeric vector}
    \item{\code{V709}}{a numeric vector}
    \item{\code{V710}}{a numeric vector}
    \item{\code{V711}}{a numeric vector}
    \item{\code{V712}}{a numeric vector}
    \item{\code{V713}}{a numeric vector}
    \item{\code{V714}}{a numeric vector}
    \item{\code{V715}}{a numeric vector}
    \item{\code{V716}}{a numeric vector}
    \item{\code{V717}}{a numeric vector}
    \item{\code{V718}}{a numeric vector}
    \item{\code{V719}}{a numeric vector}
    \item{\code{V720}}{a numeric vector}
    \item{\code{V721}}{a numeric vector}
    \item{\code{V722}}{a numeric vector}
    \item{\code{V723}}{a numeric vector}
    \item{\code{V724}}{a numeric vector}
    \item{\code{V725}}{a numeric vector}
    \item{\code{V726}}{a numeric vector}
    \item{\code{V727}}{a numeric vector}
    \item{\code{V728}}{a numeric vector}
    \item{\code{V729}}{a numeric vector}
    \item{\code{V730}}{a numeric vector}
    \item{\code{V731}}{a numeric vector}
    \item{\code{V732}}{a numeric vector}
    \item{\code{V733}}{a numeric vector}
    \item{\code{V734}}{a numeric vector}
    \item{\code{V735}}{a numeric vector}
    \item{\code{V736}}{a numeric vector}
    \item{\code{V737}}{a numeric vector}
    \item{\code{V738}}{a numeric vector}
    \item{\code{V739}}{a numeric vector}
    \item{\code{V740}}{a numeric vector}
    \item{\code{V741}}{a numeric vector}
    \item{\code{V742}}{a numeric vector}
    \item{\code{V743}}{a numeric vector}
    \item{\code{V744}}{a numeric vector}
    \item{\code{V745}}{a numeric vector}
    \item{\code{V746}}{a numeric vector}
    \item{\code{V747}}{a numeric vector}
    \item{\code{V748}}{a numeric vector}
    \item{\code{V749}}{a numeric vector}
    \item{\code{V750}}{a numeric vector}
    \item{\code{V751}}{a numeric vector}
    \item{\code{V752}}{a numeric vector}
    \item{\code{V753}}{a numeric vector}
    \item{\code{V754}}{a numeric vector}
    \item{\code{V755}}{a numeric vector}
    \item{\code{V756}}{a numeric vector}
    \item{\code{V757}}{a numeric vector}
    \item{\code{V758}}{a numeric vector}
    \item{\code{V759}}{a numeric vector}
    \item{\code{V760}}{a numeric vector}
    \item{\code{V761}}{a numeric vector}
    \item{\code{V762}}{a numeric vector}
    \item{\code{V763}}{a numeric vector}
    \item{\code{V764}}{a numeric vector}
    \item{\code{V765}}{a numeric vector}
    \item{\code{V766}}{a numeric vector}
    \item{\code{V767}}{a numeric vector}
    \item{\code{V768}}{a numeric vector}
    \item{\code{V769}}{a numeric vector}
    \item{\code{V770}}{a numeric vector}
    \item{\code{V771}}{a numeric vector}
    \item{\code{V772}}{a numeric vector}
    \item{\code{V773}}{a numeric vector}
    \item{\code{V774}}{a numeric vector}
    \item{\code{V775}}{a numeric vector}
    \item{\code{V776}}{a numeric vector}
    \item{\code{V777}}{a numeric vector}
    \item{\code{V778}}{a numeric vector}
    \item{\code{V779}}{a numeric vector}
    \item{\code{V780}}{a numeric vector}
    \item{\code{V781}}{a numeric vector}
    \item{\code{V782}}{a numeric vector}
    \item{\code{V783}}{a numeric vector}
    \item{\code{V784}}{a numeric vector}
    \item{\code{V785}}{a numeric vector}
    \item{\code{V786}}{a numeric vector}
    \item{\code{V787}}{a numeric vector}
    \item{\code{V788}}{a numeric vector}
    \item{\code{V789}}{a numeric vector}
    \item{\code{V790}}{a numeric vector}
    \item{\code{V791}}{a numeric vector}
    \item{\code{V792}}{a numeric vector}
    \item{\code{V793}}{a numeric vector}
    \item{\code{V794}}{a numeric vector}
    \item{\code{V795}}{a numeric vector}
    \item{\code{V796}}{a numeric vector}
    \item{\code{V797}}{a numeric vector}
    \item{\code{V798}}{a numeric vector}
    \item{\code{V799}}{a numeric vector}
    \item{\code{V800}}{a numeric vector}
    \item{\code{V801}}{a numeric vector}
    \item{\code{V802}}{a numeric vector}
    \item{\code{V803}}{a numeric vector}
    \item{\code{V804}}{a numeric vector}
    \item{\code{V805}}{a numeric vector}
    \item{\code{V806}}{a numeric vector}
    \item{\code{V807}}{a numeric vector}
    \item{\code{V808}}{a numeric vector}
    \item{\code{V809}}{a numeric vector}
    \item{\code{V810}}{a numeric vector}
    \item{\code{V811}}{a numeric vector}
    \item{\code{V812}}{a numeric vector}
    \item{\code{V813}}{a numeric vector}
    \item{\code{V814}}{a numeric vector}
    \item{\code{V815}}{a numeric vector}
    \item{\code{V816}}{a numeric vector}
    \item{\code{V817}}{a numeric vector}
    \item{\code{V818}}{a numeric vector}
    \item{\code{V819}}{a numeric vector}
    \item{\code{V820}}{a numeric vector}
    \item{\code{V821}}{a numeric vector}
    \item{\code{V822}}{a numeric vector}
    \item{\code{V823}}{a numeric vector}
    \item{\code{V824}}{a numeric vector}
    \item{\code{V825}}{a numeric vector}
    \item{\code{V826}}{a numeric vector}
    \item{\code{V827}}{a numeric vector}
    \item{\code{V828}}{a numeric vector}
    \item{\code{V829}}{a numeric vector}
    \item{\code{V830}}{a numeric vector}
    \item{\code{V831}}{a numeric vector}
    \item{\code{V832}}{a numeric vector}
    \item{\code{V833}}{a numeric vector}
    \item{\code{V834}}{a numeric vector}
    \item{\code{V835}}{a numeric vector}
    \item{\code{V836}}{a numeric vector}
    \item{\code{V837}}{a numeric vector}
    \item{\code{V838}}{a numeric vector}
    \item{\code{V839}}{a numeric vector}
    \item{\code{V840}}{a numeric vector}
    \item{\code{V841}}{a numeric vector}
    \item{\code{V842}}{a numeric vector}
    \item{\code{V843}}{a numeric vector}
    \item{\code{V844}}{a numeric vector}
    \item{\code{V845}}{a numeric vector}
    \item{\code{V846}}{a numeric vector}
    \item{\code{V847}}{a numeric vector}
    \item{\code{V848}}{a numeric vector}
    \item{\code{V849}}{a numeric vector}
    \item{\code{V850}}{a numeric vector}
    \item{\code{V851}}{a numeric vector}
    \item{\code{V852}}{a numeric vector}
    \item{\code{V853}}{a numeric vector}
    \item{\code{V854}}{a numeric vector}
    \item{\code{V855}}{a numeric vector}
    \item{\code{V856}}{a numeric vector}
    \item{\code{V857}}{a numeric vector}
    \item{\code{V858}}{a numeric vector}
    \item{\code{V859}}{a numeric vector}
    \item{\code{V860}}{a numeric vector}
    \item{\code{V861}}{a numeric vector}
    \item{\code{V862}}{a numeric vector}
    \item{\code{V863}}{a numeric vector}
    \item{\code{V864}}{a numeric vector}
    \item{\code{V865}}{a numeric vector}
    \item{\code{V866}}{a numeric vector}
    \item{\code{V867}}{a numeric vector}
    \item{\code{V868}}{a numeric vector}
    \item{\code{V869}}{a numeric vector}
    \item{\code{V870}}{a numeric vector}
    \item{\code{V871}}{a numeric vector}
    \item{\code{V872}}{a numeric vector}
    \item{\code{V873}}{a numeric vector}
    \item{\code{V874}}{a numeric vector}
    \item{\code{V875}}{a numeric vector}
    \item{\code{V876}}{a numeric vector}
    \item{\code{V877}}{a numeric vector}
    \item{\code{V878}}{a numeric vector}
    \item{\code{V879}}{a numeric vector}
    \item{\code{V880}}{a numeric vector}
    \item{\code{V881}}{a numeric vector}
    \item{\code{V882}}{a numeric vector}
    \item{\code{V883}}{a numeric vector}
    \item{\code{V884}}{a numeric vector}
    \item{\code{V885}}{a numeric vector}
    \item{\code{V886}}{a numeric vector}
    \item{\code{V887}}{a numeric vector}
    \item{\code{V888}}{a numeric vector}
    \item{\code{V889}}{a numeric vector}
    \item{\code{V890}}{a numeric vector}
    \item{\code{V891}}{a numeric vector}
    \item{\code{V892}}{a numeric vector}
    \item{\code{V893}}{a numeric vector}
    \item{\code{V894}}{a numeric vector}
    \item{\code{V895}}{a numeric vector}
    \item{\code{V896}}{a numeric vector}
    \item{\code{V897}}{a numeric vector}
    \item{\code{V898}}{a numeric vector}
    \item{\code{V899}}{a numeric vector}
    \item{\code{V900}}{a numeric vector}
    \item{\code{V901}}{a numeric vector}
    \item{\code{V902}}{a numeric vector}
    \item{\code{V903}}{a numeric vector}
    \item{\code{V904}}{a numeric vector}
    \item{\code{V905}}{a numeric vector}
    \item{\code{V906}}{a numeric vector}
    \item{\code{V907}}{a numeric vector}
    \item{\code{V908}}{a numeric vector}
    \item{\code{V909}}{a numeric vector}
    \item{\code{V910}}{a numeric vector}
    \item{\code{V911}}{a numeric vector}
    \item{\code{V912}}{a numeric vector}
    \item{\code{V913}}{a numeric vector}
    \item{\code{V914}}{a numeric vector}
    \item{\code{V915}}{a numeric vector}
    \item{\code{V916}}{a numeric vector}
    \item{\code{V917}}{a numeric vector}
    \item{\code{V918}}{a numeric vector}
    \item{\code{V919}}{a numeric vector}
    \item{\code{V920}}{a numeric vector}
    \item{\code{V921}}{a numeric vector}
    \item{\code{V922}}{a numeric vector}
    \item{\code{V923}}{a numeric vector}
    \item{\code{V924}}{a numeric vector}
    \item{\code{V925}}{a numeric vector}
    \item{\code{V926}}{a numeric vector}
    \item{\code{V927}}{a numeric vector}
    \item{\code{V928}}{a numeric vector}
    \item{\code{V929}}{a numeric vector}
    \item{\code{V930}}{a numeric vector}
    \item{\code{V931}}{a numeric vector}
    \item{\code{V932}}{a numeric vector}
    \item{\code{V933}}{a numeric vector}
    \item{\code{V934}}{a numeric vector}
    \item{\code{V935}}{a numeric vector}
    \item{\code{V936}}{a numeric vector}
    \item{\code{V937}}{a numeric vector}
    \item{\code{V938}}{a numeric vector}
    \item{\code{V939}}{a numeric vector}
    \item{\code{V940}}{a numeric vector}
    \item{\code{V941}}{a numeric vector}
    \item{\code{V942}}{a numeric vector}
    \item{\code{V943}}{a numeric vector}
    \item{\code{V944}}{a numeric vector}
    \item{\code{V945}}{a numeric vector}
    \item{\code{V946}}{a numeric vector}
    \item{\code{V947}}{a numeric vector}
    \item{\code{V948}}{a numeric vector}
    \item{\code{V949}}{a numeric vector}
    \item{\code{V950}}{a numeric vector}
    \item{\code{V951}}{a numeric vector}
    \item{\code{V952}}{a numeric vector}
    \item{\code{V953}}{a numeric vector}
    \item{\code{V954}}{a numeric vector}
    \item{\code{V955}}{a numeric vector}
    \item{\code{V956}}{a numeric vector}
    \item{\code{V957}}{a numeric vector}
    \item{\code{V958}}{a numeric vector}
    \item{\code{V959}}{a numeric vector}
    \item{\code{V960}}{a numeric vector}
    \item{\code{V961}}{a numeric vector}
    \item{\code{V962}}{a numeric vector}
    \item{\code{V963}}{a numeric vector}
    \item{\code{V964}}{a numeric vector}
    \item{\code{V965}}{a numeric vector}
    \item{\code{V966}}{a numeric vector}
    \item{\code{V967}}{a numeric vector}
    \item{\code{V968}}{a numeric vector}
    \item{\code{V969}}{a numeric vector}
    \item{\code{V970}}{a numeric vector}
    \item{\code{V971}}{a numeric vector}
    \item{\code{V972}}{a numeric vector}
    \item{\code{V973}}{a numeric vector}
    \item{\code{V974}}{a numeric vector}
    \item{\code{V975}}{a numeric vector}
    \item{\code{V976}}{a numeric vector}
    \item{\code{V977}}{a numeric vector}
    \item{\code{V978}}{a numeric vector}
    \item{\code{V979}}{a numeric vector}
    \item{\code{V980}}{a numeric vector}
    \item{\code{V981}}{a numeric vector}
    \item{\code{V982}}{a numeric vector}
    \item{\code{V983}}{a numeric vector}
    \item{\code{V984}}{a numeric vector}
    \item{\code{V985}}{a numeric vector}
    \item{\code{V986}}{a numeric vector}
    \item{\code{V987}}{a numeric vector}
    \item{\code{V988}}{a numeric vector}
    \item{\code{V989}}{a numeric vector}
    \item{\code{V990}}{a numeric vector}
    \item{\code{V991}}{a numeric vector}
    \item{\code{V992}}{a numeric vector}
    \item{\code{V993}}{a numeric vector}
    \item{\code{V994}}{a numeric vector}
    \item{\code{V995}}{a numeric vector}
    \item{\code{V996}}{a numeric vector}
    \item{\code{V997}}{a numeric vector}
    \item{\code{V998}}{a numeric vector}
    \item{\code{V999}}{a numeric vector}
    \item{\code{V1000}}{a numeric vector}
    \item{\code{class}}{an ordinal response with levels 0, 1, and 2}
  }
}

\examples{
data(data)
fit <- glmpath.cr(x = data[,1:1000], y = data$class)
nonzero.coef(fit)
}
\keyword{datasets}